! routing: fix `getFromDirectory` and `getFromResourceDirectory` not working properly for URIs with encoded chars
Previously the unmatched path of the request context was directly used for retrieving the file/resource content. If the path contains non-token chars that need to be encoded this results in file system/resource accesses to something like `foo%20bar.txt` instead of `foo bar.txt`. This patch fixes this problem.

The "breakingness" of this patch results from marking two internal methods private that were previously (and erroneously) public.